Enhancing cache, fix javac warning and refactor for core.
