+++ accounting/accounting/src/main/java/liquid/accounting/controller/AccountingController.java
@@ -7,7 +7,7 @@ import liquid.accounting.facade.ChargeFacade;
 import liquid.accounting.model.Charge;
 import liquid.accounting.domain.ChargeWay;
 import liquid.accounting.service.ExchangeRateService;
-import liquid.web.domain.SearchBarForm;
+import liquid.model.SearchBarForm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
