+++ accounting/accounting/src/main/java/liquid/accounting/controller/TaxRateController.java
@@ -3,7 +3,7 @@ package liquid.accounting.controller;
 import liquid.accounting.domain.TaxRate;
 import liquid.accounting.model.TaxRates;
 import liquid.accounting.service.TaxRateService;
-import liquid.model.Alert;
+import liquid.core.model.Alert;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
