+++ accounting/accounting/src/main/java/liquid/accounting/facade/ReceivableFacadeImpl.java
@@ -7,8 +7,8 @@ import liquid.accounting.service.ExchangeRateService;
 import liquid.accounting.service.InternalReceivableSummaryService;
 import liquid.order.domain.OrderEntity;
 import liquid.util.DateUtil;
-import liquid.web.domain.EnhancedPageImpl;
-import liquid.web.domain.SearchBarForm;
+import liquid.core.model.EnhancedPageImpl;
+import liquid.model.SearchBarForm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
