+++ accounting/accounting/src/main/java/liquid/accounting/restfulapi/PurchaseController.java
@@ -4,7 +4,7 @@ import liquid.operation.domain.ServiceProvider;
 import liquid.operation.service.ServiceProviderService;
 import liquid.order.domain.OrderEntity;
 import liquid.order.service.OrderService;
-import liquid.web.domain.SearchBarForm;
+import liquid.model.SearchBarForm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
