+++ container/container/src/main/java/liquid/container/controller/ContainerController.java
@@ -13,8 +13,8 @@ import liquid.operation.service.LocationService;
 import liquid.order.service.ServiceItemService;
 import liquid.transport.domain.RailContainerEntity;
 import liquid.transport.service.RailContainerService;
-import liquid.web.controller.BaseController;
-import liquid.web.domain.SearchBarForm;
+import liquid.core.controller.BaseController;
+import liquid.model.SearchBarForm;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
