+++ core/src/main/java/liquid/core/model/Alert.java
@@ -1,4 +1,4 @@
-package liquid.model;
+package liquid.core.model;
 
 import java.util.Arrays;
