+++ core/src/main/java/liquid/core/model/AlertType.java
@@ -1,4 +1,4 @@
-package liquid.model;
+package liquid.core.model;
 
 /**
  * Created by Tao Ma on 4/1/15.
