+++ core/src/main/java/liquid/core/model/EnhancedPageImpl.java
@@ -1,4 +1,4 @@
-package liquid.web.domain;
+package liquid.core.model;
 
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
@@ -8,7 +8,7 @@ import java.util.List;
 /**
  * Created by Tao Ma on 1/14/15.
  */
-public class EnhancedPageImpl<T> extends PageImpl {
+public class EnhancedPageImpl<T> extends PageImpl<T> {
     private T sum;
 
     public EnhancedPageImpl(List<T> content, Pageable pageable, long total, T sum) {
