+++ operation/operation/src/main/java/liquid/operation/controller/CustomerController.java
@@ -3,8 +3,8 @@ package liquid.operation.controller;
 import liquid.operation.domain.Customer;
 import liquid.operation.service.InternalCustomerService;
 import liquid.pinyin4j.PinyinHelper;
-import liquid.web.controller.BaseController;
-import liquid.web.domain.SearchBarForm;
+import liquid.core.controller.BaseController;
+import liquid.model.SearchBarForm;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
