+++ operation/operation/src/main/java/liquid/operation/controller/GoodsController.java
@@ -1,10 +1,10 @@
 package liquid.operation.controller;
 
-import liquid.model.Alert;
-import liquid.model.AlertType;
+import liquid.core.model.Alert;
+import liquid.core.model.AlertType;
 import liquid.operation.domain.Goods;
 import liquid.operation.service.InternalGoodsService;
-import liquid.web.controller.BaseController;
+import liquid.core.controller.BaseController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
