+++ operation/operation/src/main/java/liquid/operation/controller/LocationController.java
@@ -1,14 +1,14 @@
 package liquid.operation.controller;
 
-import liquid.domain.LocationForm;
-import liquid.model.Alert;
-import liquid.model.AlertType;
+import liquid.model.LocationForm;
+import liquid.core.model.Alert;
+import liquid.core.model.AlertType;
 import liquid.operation.domain.Location;
 import liquid.operation.domain.LocationType;
 import liquid.operation.service.InternalLocationService;
 import liquid.operation.service.LocationTypeService;
 import liquid.pinyin4j.PinyinHelper;
-import liquid.web.controller.BaseController;
+import liquid.core.controller.BaseController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
