+++ order/order/src/main/java/liquid/order/controller/OrderController.java
@@ -31,8 +31,8 @@ import liquid.transport.domain.ShipmentEntity;
 import liquid.transport.service.ShipmentService;
 import liquid.util.DateUtil;
 import liquid.validation.FormValidationResult;
-import liquid.web.controller.BaseController;
-import liquid.web.domain.SearchBarForm;
+import liquid.core.controller.BaseController;
+import liquid.model.SearchBarForm;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
