+++ order/order/src/main/java/liquid/order/restfulapi/ApiOrderController.java
@@ -4,7 +4,7 @@ import liquid.operation.domain.Customer;
 import liquid.operation.service.CustomerService;
 import liquid.order.domain.OrderEntity;
 import liquid.order.service.OrderService;
-import liquid.web.domain.SearchBarForm;
+import liquid.model.SearchBarForm;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
