+++ process/process-web/src/main/java/liquid/process/controller/TasksController.java
@@ -5,7 +5,7 @@ import liquid.process.domain.Task;
 import liquid.process.domain.TaskBar;
 import liquid.process.service.TaskService;
 import liquid.security.SecurityContext;
-import liquid.web.controller.BaseController;
+import liquid.core.controller.BaseController;
 import org.activiti.engine.ActivitiTaskAlreadyClaimedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
