+++ user/user-web/src/main/java/liquid/user/controller/AccountController.java
@@ -1,12 +1,12 @@
 package liquid.user.controller;
 
-import liquid.model.Alert;
+import liquid.core.model.Alert;
 import liquid.user.domain.GroupMember;
 import liquid.user.model.PasswordChange;
 import liquid.user.model.User;
 import liquid.user.service.UserService;
 import liquid.util.StringUtil;
-import liquid.web.controller.BaseController;
+import liquid.core.controller.BaseController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
