+++ web-console/src/main/java/liquid/config/JpaConfig.java
@@ -2,9 +2,6 @@ package liquid.config;
 
 import net.sf.ehcache.management.ManagementService;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cache.CacheManager;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.cache.ehcache.EhCacheCacheManager;
 import org.springframework.cache.ehcache.EhCacheManagerFactoryBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -26,7 +23,6 @@ import java.sql.SQLException;
 import java.util.Properties;
 
 /**
- *  
  * User: tao
  * Date: 9/24/13
  * Time: 9:06 PM
@@ -40,7 +36,6 @@ import java.util.Properties;
         "liquid.transport.repository",
         "liquid.accounting.repository"})
 @Import(PropertyPlaceholderConfig.class)
-@EnableCaching
 public class JpaConfig {
     @Value("${jdbc.url}")
     private String jdbcUrl;
@@ -122,10 +117,5 @@ public class JpaConfig {
     @Bean(initMethod = "init", destroyMethod = "dispose")
     public ManagementService managementService() {
         return new ManagementService(ehCacheManagerFactory().getObject(), mBeanServerFactory().getObject(), true, true, true, true);
-    }
-
-    @Bean
-    public CacheManager cacheManager() {
-        return new EhCacheCacheManager(ehCacheManagerFactory().getObject());
     }
 }
