+++ web-console/src/main/java/liquid/controller/OrderDistyController.java
@@ -1,12 +1,12 @@
 package liquid.controller;
 
 import liquid.domain.Disty;
-import liquid.model.Alert;
+import liquid.core.model.Alert;
 import liquid.order.domain.OrderEntity;
 import liquid.order.service.OrderService;
 import liquid.process.service.TaskService;
 import liquid.security.SecurityContext;
-import liquid.web.controller.BaseController;
+import liquid.core.controller.BaseController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
